+++ src/Service/ProductListService.php
@@ -21,6 +21,8 @@ class ProductListService
     }
 
     /**
+     * Fetches at most $maxProductCount products created the most recently.
+     *
      * @return ProductSummary[]
      */
     public function getLatestProducts(int $maxProductCount = 6) : array
@@ -30,5 +32,20 @@ class ProductListService
         }
 
         return $this->productService->search('', [], ['createdAt' => 'desc'], $maxProductCount)->getProducts();
+    }
+
+    /**
+     * Fetches at most $maxProductCount products created the most recently with the attribute $attributeId being checked.
+     *
+     * @param int $attributeId the ID of a \Wizaplace\SDK\Catalog\AttributeType::CHECKBOX_UNIQUE attribute
+     * @return ProductSummary[]
+     */
+    public function getLatestProductsWithAttributeChecked(int $attributeId, int $maxProductCount = 6): array
+    {
+        if ($maxProductCount === 0) {
+            return [];
+        }
+
+        return $this->productService->search('', [$attributeId => 'Y'], ['createdAt' => 'desc'], $maxProductCount)->getProducts();
     }
 }